+++ QRSmith/src/main/java/com/akansh/qrsmith/util/CommonShapeUtils.java
@@ -1,10 +1,10 @@
-package com.akansh.qrsmith;
+package com.akansh.qrsmith.util;
 
 import android.graphics.Canvas;
 import android.graphics.Paint;
 import android.graphics.Path;
 
-class CommonShapeUtils {
+public class CommonShapeUtils {
     public static void drawHexagon(Canvas canvas, Paint paint, float centerX, float centerY, float size) {
         Path hexagonPath = new Path();
         for (int i = 0; i < 6; i++) {
